+++ lib/ui/home/HomeAthleticsEventsWidget.dart
@@ -6,6 +6,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:illinois/model/Analytics.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Storage.dart';
+import 'package:illinois/ui/athletics/AthleticsHomePanel.dart';
 import 'package:illinois/ui/events2/Event2HomePanel.dart';
 import 'package:illinois/ui/home/HomeEvent2Widget.dart';
 import 'package:illinois/ui/home/HomePanel.dart';
@@ -81,6 +82,7 @@ class _HomeAthleticsEventsWidgetState extends State<HomeAthliticsEventsWidget> {
         updateController: widget.updateController,
         analyticsFeature: _analyticsFeature(contentType),
         emptyContentBuilder: _emptyContentBuilder(contentType),
+        onViewAll: () => _onViewAll(contentType),
         filter: _eventFilter(contentType),
         sortType: Event2SortType.dateTime,
       ),
@@ -96,6 +98,12 @@ class _HomeAthleticsEventsWidgetState extends State<HomeAthliticsEventsWidget> {
   // Analytics Feature
   AnalyticsFeature _analyticsFeature(FavoriteContentType contentType) =>
     AnalyticsFeature.Athletics;
+
+  // View All
+  void _onViewAll(FavoriteContentType contentType) {
+    Analytics().logSelect(target: "View All", source: widget.runtimeType.toString());
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsHomePanel(contentType: AthleticsContentType.events, starred: (contentType == FavoriteContentType.my),)));
+  }
 
   // Empty Content Builder
   WidgetBuilder _emptyContentBuilder(FavoriteContentType contentType) {
@@ -125,7 +133,8 @@ class _HomeAthleticsEventsWidgetState extends State<HomeAthliticsEventsWidget> {
       Uri? uri = (url != null) ? Uri.tryParse(url) : null;
       if ((uri?.scheme == localScheme) && (uri?.host == localAthleticsEventHost)) {
         Analytics().logSelect(target: 'Big 10 Events', source: runtimeType.toString());
-        Event2HomePanel.present(context, attributes: Event2HomePanel.athleticsCategoryAttributes, analyticsFeature: AnalyticsFeature.Athletics);
+        Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsHomePanel(contentType: AthleticsContentType.events, starred: true,)));
+        //Event2HomePanel.present(context, attributes: Event2HomePanel.athleticsCategoryAttributes, analyticsFeature: AnalyticsFeature.Athletics);
       }
       else if ((uri?.scheme == privacyScheme) && (uri?.host == privacyLevelHost)) {
         Analytics().logSelect(target: 'Privacy Level', source: runtimeType.toString());
@@ -133,13 +142,6 @@ class _HomeAthleticsEventsWidgetState extends State<HomeAthliticsEventsWidget> {
       }
     },
   );
-
-  // See All: Obsolete
-  // Analytics().logSelect(target: "Events");
-  // Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsHomePanel(contentType: AthleticsContentType.events)));
-
-  // Analytics().logSelect(target: "My Big 10 Events");
-  // Navigator.push(context, CupertinoPageRoute(builder: (context) => AthleticsHomePanel(contentType: AthleticsContentType.my_events)));
 }
 
 extension _FavoriteAthliticsEventsContentType on FavoriteContentType {